♻️ refactor: 에러 핸들링 함수 분리
